tech-studio-projects/financial-industry/algo-trading-platform/java/core-system/src/test/java/MarketDataProcessorTest.java
import com.system.algotrading.data.MarketDataProcessor;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertFalse;

/**
 * Unit test class for MarketDataProcessor.
 * Verifies the functionality and reliability of data processing methods.
 */
class MarketDataProcessorTest {

    private MarketDataProcessor marketDataProcessor;

    /**
     * Initializes a new instance of MarketDataProcessor before each test.
     * Ensures each test case has a fresh instance, preventing test interference.
     */
    @BeforeEach
    void setUp() {
        marketDataProcessor = new MarketDataProcessor();
    }

    /**
     * Tests the processMarketData method with sample market data.
     * Ensures that the method can handle non-empty data correctly and returns the expected processed data.
     * The test checks that the result matches the input, given the current implementation in MarketDataProcessor.
     */
    @Test
    void testProcessData() {
        // Sample data to simulate incoming market data
        String testData = "sample_market_data";

        // Process the data and retrieve the result
        String result = marketDataProcessor.processMarketData(testData);

        // Verify that the data is processed correctly by comparing the output with expected behavior
        // Expected output here is the same as input for this basic test
        assertEquals(testData, result, "Processed data should match input for this test case.");

        // Example assertion to verify non-boolean outcome (modify as needed for actual implementation)
        assertFalse(result.isEmpty(), "Processed data should not be empty.");
    }
}
